Allow value definitions with commas in them.
This makes value declarations slight more strict: there must now
be only one declaration per @value (imports still work as before,
and multiple values can be imported in a single @value).

By doing this, it allows us to be more permissive of what the
definitions can contain, thus allowing for support of things
such as:

- Multiple box-shadow shadows
- Multiple transition specifications
- Multiple media query conditions (OR'd)

It does allow for a little bit of weirder usage in this, e.g. for
specifying multiple values in a comma-separate rgba value:

@value redAndGreen: 150, 150;
.foo { background-color: rgba(redAndGreen, 0, 1); }

I don't think that's a super-desirable usage, but I also don't
think it really harms anything by being allowed, so I haven't
attempted to mitigate it.

Fixes #11.